requirements.txt
# ==================== FastAPI 核心 ====================
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# ==================== 桌面应用 ====================
pywebview>=4.0

# ==================== 环境变量管理 ====================
python-dotenv>=1.0.0

# ==================== HTTP 请求 ====================
requests>=2.31.0

# ==================== 打包工具 ====================
pyinstaller>=6.0
